requirement.txt
streamlit
tensorflow
numpy
pillow
pandas
requests
